# backend/spec/openapi.yaml
openapi: "3.1.1"
info:
  title: My Library API
  description: |
    Book library management system.
    All endpoints return JSON. Some endpoints require authentication
    (Bearer token from `/api/auth/login`).
  version: "0.3.0"
  contact:
    name: My Library
servers:
  - url: http://127.0.0.1:8000
    description: Local development
  - url: https://dingfengbo.top/my-library
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CategoryCreation:
      type: object
      required: [name]
      properties:
        name: { type: string }
        parent: { type: integer, description: "Parent category ID" }
        intro: { type: string }
    CategoryUpdate:
      type: object
      properties:
        name: { type: string }
        parent_id: { type: integer, description: "Parent category ID" }
        intro: { type: string }
    CategoryResponse:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        parent_id: { type: integer, nullable: true }
        intro: { type: string, nullable: true }
        depth: { type: integer }
        path: { type: string, nullable: true }

    # ── Bookshelf ──
...
    HTTPError:
      type: object
      properties:
        detail: { type: string }
paths:
...
  /api/categories:
    get:
      tags: [Categories]
      summary: List categories (paginated)
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 10 } }
        - { name: sort_by, in: query, schema: { type: string, default: name } }
      responses:
        "200":
          description: Paginated list
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: "#/components/schemas/CategoryResponse"
                  total_pages: { type: integer }
                  total_categories: { type: integer }
    post:
      tags: [Categories]
      summary: Create a category
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreation"
      responses:
        "200":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
  /api/categories/{category_id}:
    get:
      tags: [Categories]
      summary: Get category details
      parameters:
        - { name: category_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Category details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
    put:
      tags: [Categories]
      summary: Update a category
      security:
        - BearerAuth: []
      parameters:
        - { name: category_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
    delete:
      tags: [Categories]
      summary: Delete a category
      security:
        - BearerAuth: []
      parameters:
        - { name: category_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }

  # ═══════════════════════════════════════════════
  # Bookshelves
  # ═══════════════════════════════════════════════
...
  /api/export/:
    get:
      tags: [Export]
      summary: Export data in various formats
      security: [{ BearerAuth: [] }]
      parameters:
        - name: format
          in: query
          required: true
          schema: { type: string, enum: [sql, csv, excel, markdown, json] }
          description: Output format
        - name: scope
          in: query
          required: true
          schema: { type: string, enum: [books, authors, publishers, brands, series, categories, bookshelves, collections] }
          description: Entity type to export
      responses:
        "200":
          description: File download
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
  /api/export/sync-to-root:
    post:
      tags: [Export]
      summary: Sync user database to shared root.db
      description: |
        Syncs reference data (authors, publishers, brands, series, categories, books)
        from the current user's database to the shared root.db. Dependency order:
        reference tables first, then books, then book-author mappings.
      security: [{ BearerAuth: [] }]
      parameters:
        - name: differential
          in: query
          schema: { type: boolean, default: true }
          description: If true, only sync changed/missing entries; false for full resync
      responses:
        "200":
          description: Sync result with per-table counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  differential: { type: boolean }
                  counts:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        total: { type: integer }
                        synced: { type: integer }
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"

  # ═══════════════════════════════════════════════
  # Config
  # ═══════════════════════════════════════════════
...